Update README.md
Updated to make generic, since my particular build of Juniper Network Connect is customized for my employer just as Sean Fisk's is customized for his school.  It would be a good idea for anyone wishing to adapt this script to remap the text fields and button as referenced in the output of Sean Fisk's excellent dumpobjects.applescript implementation.